create oneof constraint and add deprecation messages in choice, also make choice extend new oneOf constraint to avoid duplicate code
